fixed bugs

+++ src/routes/devices.routes.ts
@@ -2,7 +2,7 @@ import { Request, Response, Router } from "express";
 import { HTTPS_ANSWERS } from "../utils/https-answers";
 import { deviceService } from "../services/device-service";
 import { RequestWithParams } from "../types/types";
-import { inputValidationForCookieMiddleware } from "../middleware/validation/auth-validation";
+import { checkCookieMiddleware } from "../middleware/validation/auth-validation";
 import { jwtService } from "../services/jwt-service";
 
 export const devicesRoute = Router({});
@@ -10,7 +10,7 @@ const { No_Content, OK, Not_Found, Forbidden } = HTTPS_ANSWERS;
 
 devicesRoute.get(
   "/devices",
-  inputValidationForCookieMiddleware,
+  checkCookieMiddleware,
   async (req: Request, res: Response) => {
     const id = await jwtService.getUserIdByToken(req.cookies.refreshToken);
     if (id) res.status(OK).send(await deviceService.getAllDevices(id));
@@ -19,7 +19,7 @@ devicesRoute.get(
 
 devicesRoute.delete(
   "/devices",
-  inputValidationForCookieMiddleware,
+  checkCookieMiddleware,
   async (req: Request, res: Response) => {
     const id = await jwtService.getUserIdByToken(req.cookies.refreshToken);
     if (id) await deviceService.deleteAllDevices(id);
@@ -29,11 +29,11 @@ devicesRoute.delete(
 
 devicesRoute.delete(
   "/devices/:id",
-  inputValidationForCookieMiddleware,
+  checkCookieMiddleware,
   async (req: RequestWithParams<{ id: string }>, res: Response) => {
     const userId = await jwtService.getUserIdByToken(req.cookies.refreshToken);
-    if (userId) {
-      debugger;
+    const deviceId = await deviceService.checkDeviceId(req.params.id);
+    if (userId && deviceId) {
       const result = await deviceService.deleteDevice(req.params.id, userId);
       if (result) return res.sendStatus(No_Content);
       if (!result) return res.sendStatus(Forbidden);
